// reactapp4.client/src/interfaces/MLBTeamInfo.ts
export interface MLBTeamInfo {
    gamePk: number;
    teamSide: string;
    teamId: number;
    teamName: string;
    season: string;
    springLeagueId: number | null;
    springLeagueName: string;
    springLeagueAbbreviation: string;
    allStarStatus: string;
    venueId: number | null;
    venueName: string;
    teamCode: string;
    fileCode: string;
    abbreviation: string;
    locationName: string;
    firstYearOfPlay: string;
    leagueId: number | null;
    leagueName: string;
    sportId: number | null;
    sportName: string;
    shortName: string;
    recordGamesPlayed: number | null;
    recordWildCardGamesBack: string;
    recordLeagueGamesBack: string;
    recordSpringLeagueGamesBack: string;
    recordSportGamesBack: string;
    recordDivisionGamesBack: string;
    recordConferenceGamesBack: string;
    recordLeagueRecordWins: number | null;
    recordLeagueRecordLosses: number | null;
    recordLeagueRecordTies: number | null;
    recordLeagueRecordPct: number | null;
    recordDivisionLeader: boolean | null;
    recordWins: number | null;
    recordLosses: number | null;
    recordWinningPercentage: number | null;
    franchiseName: string;
    clubName: string;
    active: boolean;
}